client side fixed

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('users', function (Blueprint $table) {
             $table->id();
-            $table->unsignedTinyInteger('region_id')->nullable();
-            $table->unsignedTinyInteger('branch_id')->nullable();
+            $table->unsignedBigInteger('region_id')->nullable();
+            $table->unsignedBigInteger('branch_id')->nullable();
