+++ database/migrations/2023_03_22_123329_create_transactions_table.php
@@ -14,11 +14,11 @@ return new class extends Migration
         Schema::create('transactions', function (Blueprint $table) {
             $table->id();
             $table->string('or')->unique();
-            $table->unsignedInteger('agent_id')->index(); // refered agent
-            $table->unsignedInteger('staff_id')->index(); // staff
-            $table->unsignedInteger('client_id')->index(); // user_id
-            $table->unsignedTinyInteger('pay_type_id')->index();
-            $table->unsignedInteger('plan_id')->index();
+            $table->unsignedBigInteger('agent_id')->index(); // refered agent
+            $table->unsignedBigInteger('staff_id')->index(); // staff
+            $table->unsignedBigInteger('client_id')->index(); // user_id
+            $table->unsignedBigInteger('pay_type_id')->index();
+            $table->unsignedBigInteger('plan_id')->index();
 
             $table->decimal('amount', 8, 2);
