+++ database/migrations/2023_08_24_063528_create_infos_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
 
             $table->unsignedBigInteger('staff_id')->index();
             $table->unsignedBigInteger('agent_id')->index();
-            $table->unsignedTinyInteger('pay_type_id')->nullable();
-            $table->unsignedTinyInteger('plan_id')->nullable();
+            $table->unsignedBigInteger('pay_type_id')->nullable();
+            $table->unsignedBigInteger('plan_id')->nullable();
 
             $table->date('bday')->nullable();
             $table->unsignedInteger('bplace_id')->nullable();
